Delete ModelingUtil and add isAny to ModelUtil
