+++ lib/draw/ArchimateRenderer.js
@@ -24,6 +24,7 @@ import {
 
 import {
   is,
+  isAny,
   getElementRef,
   getPictoRef,
   getFigureRef,
@@ -46,7 +47,6 @@ var RENDERER_IDS = new Ids();
 
 import { logger } from "../util/Logger";
 import { createLine } from 'diagram-js/lib/util/RenderUtil';
-import { isAny } from '../features/modeling/util/ModelingUtil';
 import { RELATIONSHIP_ACCESS_READ, RELATIONSHIP_ACCESS_READWRITE, RELATIONSHIP_ACCESS_WRITE } from '../metamodel/Concept';
 
 export const DEFAULT_NOTE_COLOR_RGBA = {r: '255', g: '255', b: '255', a: '100'};
@@ -391,7 +391,7 @@ export default function ArchimateRenderer(
     var color = connection.lineColor && toRgbaString(connection.lineColor) || toRgbaString(BLACK_RGBA);
     attrs = assign({
           stroke: color,
-          strokeLinecap: 'round',
+          strokeLinecap: 'none', //'round',
           strokeLinejoin: 'round',
           fill: 'none'
         }, attrs);
@@ -516,7 +516,7 @@ export default function ArchimateRenderer(
     'Access': function(parentGfx, connection) {
       var color = connection.lineColor && toRgbaString(connection.lineColor) || toRgbaString(BLACK_RGBA);
       var attrs = {
-            strokeDasharray: '1, 1',
+            strokeDasharray: '2, 2',
           };
 
       if (connection.accessType === RELATIONSHIP_ACCESS_WRITE || 
@@ -574,7 +574,7 @@ export default function ArchimateRenderer(
     'Flow': function(parentGfx, connection) {
       var color = connection.lineColor && toRgbaString(connection.lineColor) || toRgbaString(BLACK_RGBA);
       var attrs = {
-            strokeDasharray: '5, 2',
+            strokeDasharray: '6, 3',
             markerEnd: marker('closed-filled-end', color)
           };
 
@@ -584,7 +584,7 @@ export default function ArchimateRenderer(
     'Influence': function(parentGfx, connection) {
       var color = connection.lineColor && toRgbaString(connection.lineColor) || toRgbaString(BLACK_RGBA);
       var attrs = {
-            strokeDasharray: '5, 2',
+            strokeDasharray: '6, 3',
             markerEnd: marker('opened-end', color)
           };
 
@@ -595,7 +595,7 @@ export default function ArchimateRenderer(
     'Realization': function(parentGfx, connection) {
       var color = connection.lineColor && toRgbaString(connection.lineColor) || toRgbaString(BLACK_RGBA);
       var attrs = {
-            strokeDasharray: '1, 1',
+            strokeDasharray: '2, 2',
             markerEnd: marker('closed-blank-end', color)
           };
