+++ lib/draw/ArchimateRendererUtil.js
@@ -13,7 +13,7 @@ import {
   LAYER_APPLICATION, LAYER_BUSINESS, LAYER_MOTIVATION, 
   LAYER_STRATEGY, LAYER_IMP_MIG, 
   LAYER_PHYSICAL, LAYER_TECHNOLOGY,
-  ASPECT_BEHAVIOR, ASPECT_ACTIVE
+  ASPECT_BEHAVIOR, ASPECT_ACTIVE, ASPECT_PASSIVE
 } from '../metamodel/Concept';
 
 export const COLOR_LAYER_STRATEGY = {r: '245', g: '222', b: '170', a: '100'}, // 'rgba(245, 222, 170, 1)' '#F5DEAA'
@@ -39,7 +39,8 @@ const BEHAVIOR_BORDER_RADIUS = 10,
 
 export const BORDER_ASPECT_MAP = new Map([
   [ASPECT_BEHAVIOR, BEHAVIOR_BORDER_RADIUS],
-  [ASPECT_ACTIVE, ACTIVE_BORDER_RADIUS]
+  [ASPECT_ACTIVE, ACTIVE_BORDER_RADIUS],
+  [ASPECT_PASSIVE, ACTIVE_BORDER_RADIUS]
 ]);
 
 export const LIGHT_GREY_RGBA = {r: '211', g: '211', b: '211', a: '100'}; // 'rgba(211, 211, 211, 1)';
