+++ lib/features/auto-place/ArchimateAutoPlaceUtil.js
@@ -1,5 +1,4 @@
-import { is } from '../../util/ModelUtil';
-import { isAny } from '../modeling/util/ModelingUtil';
+import { is, isAny } from '../../util/ModelUtil';
 
 import {
   getMid,
