+++ lib/features/canvas-create/CanvasCreate.js
@@ -7,6 +7,7 @@ import {
 } from 'min-dash';
 
 import {
+  isAny,
   getViewElement
 } from '../../util/ModelUtil';
 
@@ -15,7 +16,6 @@ import COLORS from '../../util/ColorUtil';
 import {
   toPoint
 } from 'diagram-js/lib/util/Event';
-import { isAny } from '../modeling/util/ModelingUtil';
 
 import { logger } from "../../util/Logger";
