+++ lib/features/context-pad/ContextPadProvider.js
@@ -10,8 +10,7 @@ import {
 } from 'diagram-js/lib/util/Mouse';
 
 import COLORS from '../../util/ColorUtil';
-import { is, getViewElement, getElementRef } from '../../util/ModelUtil';
-import { isAny } from '../modeling/util/ModelingUtil';
+import { is, isAny, getViewElement, getElementRef } from '../../util/ModelUtil';
 
 import { logger } from "../../util/Logger";
