+++ lib/features/label-editing/LabelEditingPreview.js
@@ -2,9 +2,7 @@ import {
   remove as svgRemove
 } from 'tiny-svg';
 
-import { isAny } from '../modeling/util/ModelingUtil';
-
-import { getViewElement } from '../../util/ModelUtil';
+import { isAny, getViewElement } from '../../util/ModelUtil';
 
 var MARKER_HIDDEN = 'djs-element-hidden',
     MARKER_LABEL_HIDDEN = 'djs-label-hidden';
