+++ lib/features/label-editing/LabelEditingProvider.js
@@ -8,9 +8,7 @@ import {
 
 import { logger } from "../../util/Logger";
 
-import { getViewElement } from '../../util/ModelUtil';
-
-import { isAny } from '../modeling/util/ModelingUtil';
+import { isAny, getViewElement } from '../../util/ModelUtil';
 
 import {
   getExternalLabelMid,
