+++ lib/features/label-editing/LabelUtil.js
@@ -1,5 +1,4 @@
-import { isAny } from '../modeling/util/ModelingUtil';
-import { is, getElementRef } from '../../util/ModelUtil';
+import { is, isAny, getElementRef } from '../../util/ModelUtil';
 
 function getLabelAttr(elementRef) {
   if (isAny(elementRef, [ 'archimate:BaseElement', 'archimate:Group' ])) {
