+++ lib/features/modeling/ArchimateFactory.js
@@ -4,11 +4,7 @@ import {
 } from 'min-dash';
 
 import {
-  isAny
-} from './util/ModelingUtil';
-
-import {
-  is,
+  is
 } from '../../util/ModelUtil';
 
 import { logger } from "../../util/Logger";
@@ -19,11 +15,8 @@ export default function ArchimateFactory(moddle) {
 
 ArchimateFactory.$inject = [ 'moddle' ];
 
-
 ArchimateFactory.prototype._needsId = function(element) {
-  return isAny(element, [
-    'archimate:IdObject'
-  ]);
+  return is(element, 'archimate:IdObject');
 };
 
 ArchimateFactory.prototype._ensureId = function(element) {
