+++ lib/features/modeling/ConnectionUpdater.js
@@ -15,6 +15,7 @@ import {
 
 import {
   is,
+  isAny,
   getViewElement,
 } from '../../util/ModelUtil';
 
@@ -23,7 +24,6 @@ import { logger } from "../../util/Logger";
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
 import { COLOR_LAYER_MAP, BLACK_SHADOW_RGBA } from '../../draw/ArchimateRendererUtil';
 import { DEFAULT_NOTE_COLOR_RGBA } from '../../draw/ArchimateRenderer';
-import { isAny } from './util/ModelingUtil';
 import { RELATIONSHIP_ACCESS, RELATIONSHIP_ASSOCIATION, RELATIONSHIP_INFLUENCE } from '../../metamodel/Concept';
 
 /**
