+++ lib/features/ordering/ArchimateOrderingProvider.js
@@ -3,16 +3,12 @@ import inherits from 'inherits-browser';
 import OrderingProvider from 'diagram-js/lib/features/ordering/OrderingProvider';
 
 import {
-  isAny
-} from '../modeling/util/ModelingUtil';
-
-import {
   findIndex,
   find
 } from 'min-dash';
 
 import { logger } from "../../util/Logger";
-import { is } from '../../util/ModelUtil';
+import { is, isAny } from '../../util/ModelUtil';
 
 /**
  * a simple ordering provider that makes sure:
