+++ lib/features/palette/PaletteProvider.js
@@ -185,10 +185,10 @@ PaletteProvider.prototype.getPaletteEntries = function(element) {
       separator: true
     }
   });
-  
+
   assign(actions, {
     'create.note': createAction(
-      'archimate:Note', 'Other', 'archimate-tool-note',
+      'archimate:Note', 'Other', 'archimate-menu-tool-note',
       translate('Create Note '))
   });
 /*
