+++ lib/features/popup-menu/ConnectionMenuProvider.js
@@ -201,7 +201,7 @@ ConnectionMenuProvider.prototype._getAccessType = function(element) {
     }
   }
 
-  var accessType = element.accesType;
+  var accessType = element.accessType;
   var isRead = accessType === RELATIONSHIP_ACCESS_READ,
       isReadWrite = accessType === RELATIONSHIP_ACCESS_READWRITE,
       isWrite = accessType === RELATIONSHIP_ACCESS_WRITE;
@@ -214,7 +214,7 @@ ConnectionMenuProvider.prototype._getAccessType = function(element) {
       active: isRead,
       action: toggleAccessType,
       options: {
-        accesType: RELATIONSHIP_ACCESS_READ
+        accessType: RELATIONSHIP_ACCESS_READ
       }
     },
     {
@@ -224,7 +224,7 @@ ConnectionMenuProvider.prototype._getAccessType = function(element) {
       active: isWrite,
       action: toggleAccessType,
       options: {
-        accesType: RELATIONSHIP_ACCESS_WRITE
+        accessType: RELATIONSHIP_ACCESS_WRITE
       }
     },
     {
@@ -234,7 +234,7 @@ ConnectionMenuProvider.prototype._getAccessType = function(element) {
       active: isReadWrite,
       action: toggleAccessType,
       options: {
-        accesType: RELATIONSHIP_ACCESS_READWRITE
+        accessType: RELATIONSHIP_ACCESS_READWRITE
       }
     }
   ];
