+++ lib/features/popup-menu/ConnectionOptions.js
@@ -20,11 +20,11 @@ const s = RELATIONSHIP_SPECIALIZATION,
 
 const groupDirect = {
           id: 'group-direct',
-          name: 'Direct connection'
+          name: 'New direct connection'
         },
       groupReverse = {
           id: 'group-reverse',
-          name: 'Reverse connection'
+          name: 'New reverse connection'
       };
 
 const RELATIONSHIP_MENU_MAP = new Map([
