+++ lib/features/rules/ArchimateRules.js
@@ -5,17 +5,15 @@ import {
 import inherits from 'inherits-browser';
 
 import {
-  is, getElementRef, getViewElement, VIEW } from '../../util/ModelUtil';
+  is, isAny, getElementRef, getViewElement, VIEW } from '../../util/ModelUtil';
 
 import {
   isLabel
 } from '../../util/LabelUtil';
 
 import RuleProvider from 'diagram-js/lib/features/rules/RuleProvider';
-import { isAny } from '../modeling/util/ModelingUtil';
 
 import { logger } from "../../util/Logger";
-import { RELATIONSHIP_ASSIGNMENT } from '../../metamodel/Concept';
 import { isRelationshipAllowed } from '../popup-menu/ConnectionOptions';
 
 /**
