+++ lib/import/ArchimateImporter.js
@@ -2,7 +2,7 @@ import {
   assign, map
 } from 'min-dash';
 
-import { is } from '../util/ModelUtil';
+import { is, isAny } from '../util/ModelUtil';
 
 import {
   isLabelExternal,
@@ -19,7 +19,6 @@ import {
 
 import { logger } from "../util/Logger";
 import connectionPreview from 'diagram-js/lib/features/connection-preview';
-import { isAny } from '../features/modeling/util/ModelingUtil';
 
 function elementData(semantic, attrs) {
   return assign({
