+++ lib/util/ModelUtil.js
@@ -1,3 +1,7 @@
+import {
+  some
+} from 'min-dash';
+
 import * as CX from "../metamodel/Concept";
 import * as RX from "../metamodel/RelationshipRelationshipMap";
 import * as BX from "../metamodel/BusinessRelationshipMap";
@@ -204,4 +208,42 @@ export function isViewElementType(element, type) {
   return (element && element.businessObject) || null;
 
     // return (element && element.diObject) || element;
+}
+
+/**
+ * Return true if element has any of the given types.
+ *
+ * @param {djs.model.Base} element
+ * @param {Array<String>} types
+ *
+ * @return {Boolean}
+ */
+export function isAny(element, types) {
+  return some(types, function(t) {
+    return is(element, t);
+  });
+}
+
+
+/**
+ * Return the parent of the element with any of the given types.
+ *
+ * @param {djs.model.Base} element
+ * @param {String|Array<String>} anyType
+ *
+ * @return {djs.model.Base}
+ */
+export function getParent(element, anyType) {
+
+  if (typeof anyType === 'string') {
+    anyType = [ anyType ];
+  }
+
+  while ((element = element.parent)) {
+    if (isAny(element, anyType)) {
+      return element;
+    }
+  }
+
+  return null;
 }
